requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
amqp==5.0.2               # via kombu
apipkg==1.5               # via execnet
asgiref==3.3.1            # via django
attrs==20.3.0             # via pytest
autopep8==1.5.4           # via django-silk
backcall==0.2.0           # via ipython
billiard==3.6.3.0         # via celery
boto3==1.16.18            # via -r requirements.in
botocore==1.19.18         # via boto3, s3transfer
celery==5.0.2             # via -r requirements.in, django-celery-email
certifi==2020.11.8        # via requests
cffi==1.14.3              # via cryptography
chardet==3.0.4            # via requests
click-didyoumean==0.0.3   # via celery
click-repl==0.1.6         # via celery
click==7.1.2              # via celery, click-didyoumean, click-repl
coverage==5.3             # via pytest-cov
cryptography==3.2.1       # via -r requirements.in, pyjwt, pyopenssl
decorator==4.4.2          # via ipython
defusedxml==0.6.0         # via python3-openid
dj-database-url==0.5.0    # via -r requirements.in
django-allauth==0.43.0    # via -r requirements.in
django-anymail[mailgun]==8.1  # via -r requirements.in
django-appconf==1.0.4     # via django-celery-email
django-celery-email==3.0.0  # via -r requirements.in
django-environ==0.4.5     # via -r requirements.in
django-extensions==3.0.9  # via -r requirements.in
django-model-utils==4.0.0  # via -r requirements.in
django-redis==4.12.1      # via -r requirements.in
django-silk==4.1.0        # via -r requirements.in
django-storages==1.10.1   # via -r requirements.in
django-widget-tweaks==1.4.8  # via -r requirements.in
django==3.1.3             # via -r requirements.in, django-allauth, django-anymail, django-appconf, django-celery-email, django-model-utils, django-redis, django-silk, django-storages
execnet==1.7.1            # via pytest-xdist
factory-boy==3.1.0        # via -r requirements.in
faker==4.14.2             # via factory-boy
future==0.18.2            # via -r requirements.in
gprof2dot==2019.11.30     # via django-silk
gunicorn==20.0.4          # via -r requirements.in
idna==2.10                # via requests
iniconfig==1.1.1          # via pytest
ipython-genutils==0.2.0   # via traitlets
ipython==7.19.0           # via -r requirements.in
jedi==0.17.2              # via ipython
jinja2==2.11.2            # via django-silk
jmespath==0.10.0          # via boto3, botocore
kombu==5.0.2              # via celery
markupsafe==1.1.1         # via jinja2
oauthlib==3.1.0           # via requests-oauthlib
packaging==20.4           # via pytest
parso==0.7.1              # via jedi
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==8.0.1             # via -r requirements.in
pluggy==0.13.1            # via pytest
prompt-toolkit==3.0.8     # via click-repl, ipython
psycopg2-binary==2.8.6    # via -r requirements.in
ptyprocess==0.6.0         # via pexpect
py==1.9.0                 # via pytest, pytest-forked
pycodestyle==2.6.0        # via autopep8
pycparser==2.20           # via cffi
pygments==2.7.2           # via django-silk, ipython
pyjwt[crypto]==1.7.1      # via django-allauth
pyopenssl==19.1.0         # via -r requirements.in
pyparsing==2.4.7          # via packaging
pytest-cov==2.10.1        # via -r requirements.in
pytest-django==4.1.0      # via -r requirements.in
pytest-forked==1.3.0      # via pytest-xdist
pytest-mock==3.3.1        # via -r requirements.in
pytest-randomly==3.4.1    # via -r requirements.in
pytest-xdist==2.1.0       # via -r requirements.in
pytest==6.1.2             # via -r requirements.in, pytest-cov, pytest-django, pytest-forked, pytest-mock, pytest-randomly, pytest-xdist
python-dateutil==2.8.1    # via botocore, django-silk, faker
python3-openid==3.2.0     # via django-allauth
pytz==2020.4              # via -r requirements.in, celery, django, django-silk
redis==3.5.3              # via -r requirements.in, django-redis
requests-oauthlib==1.3.0  # via django-allauth
requests==2.25.0          # via django-allauth, django-anymail, django-silk, requests-oauthlib
s3transfer==0.3.3         # via boto3
six==1.15.0               # via click-repl, cryptography, packaging, pyopenssl, python-dateutil
sorl-thumbnail==12.6.3    # via -r requirements.in
sqlparse==0.4.1           # via django, django-silk
text-unidecode==1.3       # via faker
toml==0.10.2              # via autopep8, pytest
traitlets==5.0.5          # via ipython
unidecode==1.1.1          # via -r requirements.in
urllib3==1.26.2           # via botocore, requests
vine==5.0.0               # via amqp, celery
wcwidth==0.2.5            # via prompt-toolkit
werkzeug==1.0.1           # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
